+++ src/modules/FSXARemoteApi.ts
@@ -14,9 +14,9 @@ import {
   NavigationItemFilter,
   RemoteApiFilterOptions,
   MappedCaasItem,
-  RemoteProjectConfiguration,
   SortParams,
   CaasApi_Item,
+  RemoteProjectConfiguration,
   NormalizedFetchResponse,
   NormalizedProjectPropertyResponse,
 } from '../types'
@@ -53,7 +53,7 @@ export class FSXARemoteApi implements FSXAApi {
   private _navigationServiceURL: string = this.navigationServiceURL
   private _tenantID: string = this.tenantID
   private _projectID: string = this.projectID
-  private _remotes: Record<string, { id: string; locale: string }> = this.remotes
+  private _remotes: RemoteProjectConfiguration = this.remotes
   private _contentMode: 'preview' | 'release' = this.contentMode
   private _maxReferenceDepth?: number
   private _customMapper?: CustomMapper
@@ -109,7 +109,6 @@ export class FSXARemoteApi implements FSXAApi {
     this._caasItemFilter = filterOptions?.caasItemFilter
 
     this._logger.debug('FSXARemoteApi created', {
-      apikey,
       caasURL,
       navigationServiceURL,
       tenantID,
@@ -133,11 +132,13 @@ export class FSXARemoteApi implements FSXAApi {
   }
 
   private getRemoteProject(remoteProject: string) {
-    const projectId = this.remotes[remoteProject]?.id
-    if (!projectId) {
+    const remoteProjectConfig = Object.values(this._remotes)
+    const foundRemoteProject = remoteProjectConfig.find((config) => config.id === remoteProject)
+    if (!foundRemoteProject) {
       throw new Error(FSXAApiErrors.UNKNOWN_REMOTE)
     }
-    return projectId
+
+    return remoteProject
   }
 
   /**
@@ -841,7 +842,7 @@ export class FSXARemoteApi implements FSXAApi {
   /**
    * @returns the configured remote project configuration
    */
-  public get remotes(): Record<string, { id: string; locale: string }> {
+  public get remotes(): RemoteProjectConfiguration {
     return this._remotes
   }
 
@@ -858,7 +859,7 @@ export class FSXARemoteApi implements FSXAApi {
     }
    ```
    */
-  public set remotes(value: Record<string, { id: string; locale: string }>) {
+  public set remotes(value: RemoteProjectConfiguration) {
     const keys = Object.keys(value)
     keys.forEach((key) => {
       const { id, locale } = value[key]